+++ files/gitlab-cookbooks/gitlab/recipes/gitlab-shell.rb
@@ -91,7 +91,7 @@ templatesymlink "Create a config.yml and create a symlink to Rails root" do
   link_from File.join(gitlab_shell_dir, "config.yml")
   link_to File.join(gitlab_shell_var_dir, "config.yml")
   source "gitlab-shell-config.yml.erb"
-  variables( {
+  variables({
     :user => git_user,
     :api_url => api_url,
     :authorized_keys => authorized_keys,
@@ -105,7 +105,8 @@ templatesymlink "Create a config.yml and create a symlink to Rails root" do
     :log_level => node['gitlab']['gitlab-shell']['log_level'],
     :audit_usernames => node['gitlab']['gitlab-shell']['audit_usernames'],
     :http_settings => node['gitlab']['gitlab-shell']['http_settings'],
-    :git_annex_enabled => node['gitlab']['gitlab-shell']['git_annex_enabled']
+    :git_annex_enabled => node['gitlab']['gitlab-shell']['git_annex_enabled'],
+    :git_trace_log_file => File.join(log_directory, 'gitlab-shell-git-trace.log')
   })
 end